Add --search option
Sometimes when you have a big report with lots of hosts, you want to search within those
hosts.

This adds a `--search` option which, for each row in our output, it
searches the key/vals for the input string, and if it matches, displays
the table for that host, otherwise, the host won't even render.